keep wrapPreAndFinallyExecutionSqlQuery in snowflake module
