+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-snowflake/legend-engine-xt-relationalStore-snowflake-execution/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/test/semiStructured/TestSnowflakeSemiStructuredFlattening.java
@@ -20,7 +20,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertThrows;
 
-public class TestSnowflakeSemiStructuredFlattening extends AbstractTestSemiStructured
+public class TestSnowflakeSemiStructuredFlattening extends AbstractTestSnowflakeSemiStructured
 {
     private static final String snowflakeMapping = "flatten::mapping::SnowflakeMapping";
     private static final String snowflakeRuntime = "flatten::runtime::SnowflakeRuntime";
