+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-snowflake/legend-engine-xt-relationalStore-snowflake-execution/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/test/semiStructured/TestSnowflakeSemiStructuredMultiBindingMapping.java
@@ -17,7 +17,7 @@ package org.finos.legend.engine.plan.execution.stores.relational.test.semiStruct
 import org.junit.Assert;
 import org.junit.Test;
 
-public class TestSnowflakeSemiStructuredMultiBindingMapping extends AbstractTestSemiStructured
+public class TestSnowflakeSemiStructuredMultiBindingMapping extends AbstractTestSnowflakeSemiStructured
 {
     private static final String snowflakeMapping = "multiBinding::mapping::SnowflakeMapping";
     private static final String snowflakeRuntime = "multiBinding::runtime::SnowflakeRuntime";
