+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-snowflake/legend-engine-xt-relationalStore-snowflake-execution/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/test/semiStructured/TestSnowflakeSemiStructuredParseJsonMapping.java
@@ -17,7 +17,7 @@ package org.finos.legend.engine.plan.execution.stores.relational.test.semiStruct
 import org.junit.Assert;
 import org.junit.Test;
 
-public class TestSnowflakeSemiStructuredParseJsonMapping extends AbstractTestSemiStructured
+public class TestSnowflakeSemiStructuredParseJsonMapping extends AbstractTestSnowflakeSemiStructured
 {
     private static final String snowflakeMapping = "parseJson::mapping::SnowflakeMapping";
     private static final String snowflakeRuntime = "parseJson::runtime::SnowflakeRuntime";
